src/backofficer-func/requirements.txt
applicationinsights
azure-functions
opencensus-ext-azure
requests
pyjwt
